Ditching Vagrant for AWS
Vagrant is cool, but paying for a server that can run a decent-sized vm but not using
that capacity most of the time is pretty pointless. Spinning up an AWS instance on
demand makes much more sense. Updated readme to reflect that shift in planned architecture.

Also added an application.yml.example to stand in for the one that's gitignored